backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
textstat==0.7.3
pymupdf==1.23.8
python-docx==1.1.0
pydantic==2.5.2
groq>=0.9.0
python-dotenv>=1.0.0